camera: Replace testcamera.c with testcameraminimal.c
